requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

openai-whisper==20230314
pydub==0.25.1
